requirements.txt
maafw
numpy
